Remove rems from font-size mixin
Since rems are output second, it occasionally causes
rounding errors when a decimal is used as the basis
for another property such as line-height, leading to
imperfect values such as 17.9999 when multiplying the
rems by 1.5

Furthermore, modern browsers have no problem scaling
their entire interface even when pixel values are used
as a font-size unit. There's no loss in functionality
by sticking to px-only as the mixin's output.